refactor: improve service structure
