+++ src/Application/Services/Hasher/HasherService.cs
@@ -1,5 +1,5 @@
 using Domain.Entities;
-using Domain.Services;
+using Domain.Services.Hasher;
 using Microsoft.AspNetCore.Identity;
 
 namespace Application.Services.Hasher;
