--- src/Domain/Services/IJwtService.cs
@@ -1,9 +1,0 @@
-using Domain.Entities;
-
-namespace Domain.Services;
-
-public interface IJwtService
-{
-  string GenerateToken(User user);
-  int? GetUserIdFromToken();
-}
