+++ src/Domain/Services/Jwt/IJwtService.cs
@@ -1,0 +1,9 @@
+using Domain.Entities;
+
+namespace Domain.Services.Jwt;
+
+public interface IJwtService
+{
+  string Generate(User user);
+  int GetCurrentUser();
+}
